RPG Game with SFML
This is a story game. Our character Kamil Goldlover goes on a magical journey to save his little sister and encounters various difficulties. Our character, who fights with monsters and talks to the guard and headman on the island, reaches the ballot box at the end of the story and makes a choice. In this text-heavy game, the choices the character makes change the ending of the story.